requirements.txt
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
Flask-Migrate==4.0.5
openai>=1.30.0
pytrends==4.9.2
APScheduler==3.10.4
python-dotenv==1.0.0
requests==2.31.0
beautifulsoup4==4.12.3
Markdown==3.5.2
gunicorn==21.2.0
